requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
pyrogram==1.4.9
youtube-search-python
yt-dlp
lyricsgenius
speedtest-cli
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
pillow
motor
psutil
future
wget
